Make app config tests run without Flask
